+++ src/app/admin/components/admin-product-form/admin-product-form.component.ts
@@ -1,4 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
+import { pluck } from 'rxjs/operators';
+import { Product } from 'src/app/products/models/product.model';
+import { ProductsService } from 'src/app/products/services/products.service';
 
 @Component({
   selector: 'app-admin-product-form',
@@ -6,10 +10,24 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./admin-product-form.component.scss']
 })
 export class AdminProductFormComponent implements OnInit {
-
-  constructor() { }
+  product: Product;
+  
+  constructor(
+    private productsService: ProductsService,
+    private route: ActivatedRoute,
+    private router: Router
+  ) { }
 
   ngOnInit(): void {
+    this.route.data.pipe(pluck('product')).subscribe((product: Product) => {
+      this.product = { ...product };
+    });
   }
 
+  onSaveProduct() {
+  }
+
+  onGoBack() {
+    this.router.navigate(['/admin/products']);
+  }
 }