+++ src/app/admin/components/admin-products-list/admin-products-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Observable } from 'rxjs';
 import { Product } from 'src/app/products/models/product.model';
 import { ProductsService } from 'src/app/products/services/products.service';
 
@@ -8,7 +9,7 @@ import { ProductsService } from 'src/app/products/services/products.service';
   styleUrls: ['./admin-products-list.component.scss']
 })
 export class AdminProductsListComponent implements OnInit {
-  products: Promise<Array<Product>>;
+  products: Observable<Array<Product>>;
 
   constructor(
     private productsService: ProductsService,