+++ src/app/admin/admin.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { SharedModule } from '../shared/shared.module';
 import { AdminRoutingModule } from './admin-routing.module';
 import { ProductsModule } from '../products/products.module';
-import { AdminProductsListComponent } from './components/admin-products-list/admin-products-list.component';
 import { AdminComponent } from './admin.component';
 import { AdminProductFormComponent } from './components/admin-product-form/admin-product-form.component';
 import { AdminOrdersListComponent } from './components/admin-orders-list/admin-orders-list.component';