+++ changelog.md
@@ -55,3 +55,4 @@ Created ProcessOrder component; Added IsEmptyCart guard
 
 02.03.2021
 Improved usability
+Added admin routes and stub components